Initial revision.
This is the Python command line tester for the CLS REST API. It makes
each type of API call and verifies that the JSON Object returned is
what’s expected. Changes to the implementation will cause this code to
break, which is intended, so that I can make sure I don’t break
backwards compatibility.